Extract workspace-picture code into dedicated files
Pure move plus one guard: sidebar picture menu/avatar UI moves from
ContentView.swift to Sources/Sidebar/SidebarWorkspacePictureUI.swift,
workspace picture mutations from Workspace.swift to
Sources/Workspace+Picture.swift (the inline restore migration becomes
restorePicture(fromSnapshotWorkspaceId:)), and the mobile fetch/cache
from MobileShellComposite.swift to
MobileShellComposite+WorkspacePictures.swift. The mobile extension also
enforces the 512KB avatar cap client-side so an oversized RPC response
is dropped instead of cached.
