+++ Packages/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite.swift
@@ -208,7 +208,9 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
     private let identityProvider: (any MobileIdentityProviding)?
     private let reachability: any ReachabilityProviding
     private let pairingHintDefaults: UserDefaults
-    private let clientID: String
+    /// Internal (not private) so the workspace-pictures extension file can
+    /// build RPC params with it.
+    let clientID: String
     /// The injected, fire-and-forget product-analytics emitter. Defaults to
     /// ``NoopAnalytics`` so previews/tests inject nothing.
     private let analytics: any AnalyticsEmitting
@@ -241,7 +243,9 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
     /// `public` so the DEV feedback-submit affordance can ``DiagnosticLog/export()``
     /// it.
     public let diagnosticLog: DiagnosticLog?
-    private var remoteClient: MobileCoreRPCClient? {
+    /// Internal (not private) so the workspace-pictures extension file can
+    /// guard fetch results against reconnects.
+    var remoteClient: MobileCoreRPCClient? {
         didSet {
             if remoteClient == nil {
                 stopTerminalRefreshPolling()
@@ -273,12 +277,12 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
     /// Mac changes the hash whenever the picture changes, so a stale avatar is
     /// never served. Bounded by `maxCachedWorkspacePictures` so a long-lived
     /// session with many edited avatars can't grow it without limit.
-    private var workspacePictureBytesByHash: [String: Data] = [:]
+    var workspacePictureBytesByHash: [String: Data] = [:]
     /// LRU order of cached picture hashes (oldest first) for bounded eviction.
-    private var workspacePictureHashLRU: [String] = []
+    var workspacePictureHashLRU: [String] = []
     /// Hashes with an in-flight fetch, so overlapping list refreshes don't kick
     /// off duplicate picture requests for the same avatar.
-    private var workspacePictureFetchInFlight: Set<String> = []
+    var workspacePictureFetchInFlight: Set<String> = []
     nonisolated static let maxCachedWorkspacePictures = 128
     private var createWorkspaceTaskID: UUID?
     private var createTerminalTaskID: UUID?
@@ -3008,106 +3012,6 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
                 return terminal
             }
             return workspace
-        }
-    }
-
-    /// Kick off fetches for any workspace picture whose hash is not yet cached.
-    /// Idempotent and de-duplicated: each hash fetches at most once. The bytes are
-    /// fetched on demand (not in the list payload) so `workspace.updated` stays
-    /// small and an unchanged avatar is never re-sent.
-    private func fetchMissingWorkspacePictures() {
-        guard let client = remoteClient else { return }
-        var pending: [(workspaceID: String, hash: String)] = []
-        var seenHashes: Set<String> = []
-        for workspace in workspaces {
-            guard let hash = workspace.pictureHash,
-                  workspacePictureBytesByHash[hash] == nil,
-                  !workspacePictureFetchInFlight.contains(hash),
-                  seenHashes.insert(hash).inserted else {
-                continue
-            }
-            pending.append((workspace.id.rawValue, hash))
-        }
-        guard !pending.isEmpty else { return }
-        for entry in pending {
-            workspacePictureFetchInFlight.insert(entry.hash)
-        }
-        Task { @MainActor [weak self] in
-            for entry in pending {
-                await self?.fetchWorkspacePicture(
-                    client: client,
-                    workspaceID: entry.workspaceID,
-                    hash: entry.hash
-                )
-            }
-        }
-    }
-
-    private func fetchWorkspacePicture(
-        client: MobileCoreRPCClient,
-        workspaceID: String,
-        hash: String
-    ) async {
-        defer { workspacePictureFetchInFlight.remove(hash) }
-        do {
-            let request = try MobileCoreRPCClient.requestData(
-                method: "mobile.workspace.picture.get",
-                params: [
-                    "workspace_id": workspaceID,
-                    "hash": hash,
-                    "client_id": clientID,
-                ]
-            )
-            let data = try await client.sendRequest(request)
-            guard self.remoteClient === client else { return }
-            let response = try MobileWorkspacePictureResponse.decode(data)
-            guard response.hash == hash, let imageData = response.imageData else { return }
-            cacheWorkspacePicture(imageData, forHash: hash)
-            applyCachedWorkspacePictures()
-        } catch {
-            mobileShellLog.info("workspace picture fetch failed hash=\(hash, privacy: .public) error=\(String(describing: error), privacy: .private)")
-        }
-    }
-
-    private func cacheWorkspacePicture(_ data: Data, forHash hash: String) {
-        if workspacePictureBytesByHash[hash] == nil {
-            workspacePictureHashLRU.append(hash)
-        }
-        workspacePictureBytesByHash[hash] = data
-        guard workspacePictureHashLRU.count > Self.maxCachedWorkspacePictures else { return }
-        // Single bounded pass, oldest first: evict unreferenced hashes until the
-        // cache fits. A hash still referenced by a live workspace is never
-        // evicted, so when every cached avatar is live the cache simply stays
-        // over budget for this round (bounded by the live workspace count).
-        let liveHashes = Set(workspaces.compactMap(\.pictureHash))
-        var overflow = workspacePictureHashLRU.count - Self.maxCachedWorkspacePictures
-        var retained: [String] = []
-        retained.reserveCapacity(workspacePictureHashLRU.count)
-        for candidate in workspacePictureHashLRU {
-            if overflow > 0, !liveHashes.contains(candidate) {
-                workspacePictureBytesByHash.removeValue(forKey: candidate)
-                overflow -= 1
-            } else {
-                retained.append(candidate)
-            }
-        }
-        workspacePictureHashLRU = retained
-    }
-
-    /// Re-stamp `pictureData` on each workspace from the cache (after a fetch
-    /// resolved new bytes), publishing the change so rows pick up the avatar.
-    private func applyCachedWorkspacePictures() {
-        var didChange = false
-        let updated = workspaces.map { workspace -> MobileWorkspacePreview in
-            let resolved = workspace.pictureHash.flatMap { workspacePictureBytesByHash[$0] }
-            guard workspace.pictureData != resolved else { return workspace }
-            var copy = workspace
-            copy.pictureData = resolved
-            didChange = true
-            return copy
-        }
-        if didChange {
-            workspaces = updated
         }
     }
