+++ Sources/Workspace.swift
@@ -325,19 +325,7 @@ extension Workspace {
         isPinned = snapshot.isPinned
         groupId = snapshot.groupId
 
-        // The picture file is named by workspace id, but restore rebuilds this
-        // workspace under a fresh UUID, so re-home the original id's picture onto
-        // the new id. Fall back to re-deriving from disk so a snapshot hash with
-        // no backing file resolves to no avatar instead of a dangling hash.
-        if let originalWorkspaceId = snapshot.workspaceId,
-           let migratedHash = WorkspacePictureStore.shared.migratePicture(
-               from: originalWorkspaceId,
-               to: id
-           ) {
-            pictureHash = migratedHash
-        } else {
-            reloadPictureHashFromStore()
-        }
+        restorePicture(fromSnapshotWorkspaceId: snapshot.workspaceId)
 
         // Status entries and agent PIDs are ephemeral runtime state tied to running
         // processes (e.g. claude_code "Running"). Don't restore them across app
@@ -12355,32 +12343,6 @@ final class Workspace: Identifiable, ObservableObject {
         } else {
             customColor = nil
         }
-    }
-
-    /// Set this workspace's picture from an image, persisting the downscaled PNG
-    /// via `WorkspacePictureStore` and publishing the new content hash. Returns
-    /// `false` when the image can't be encoded. The `@Published` hash change is
-    /// what drives the sidebar avatar and the mobile-list push.
-    @discardableResult
-    func setPicture(_ image: NSImage) -> Bool {
-        guard let hash = WorkspacePictureStore.shared.setPicture(image, for: id) else {
-            return false
-        }
-        pictureHash = hash
-        return true
-    }
-
-    /// Remove this workspace's picture from disk and clear the published hash.
-    func clearPicture() {
-        WorkspacePictureStore.shared.removePicture(for: id)
-        pictureHash = nil
-    }
-
-    /// Re-derive the published picture hash from disk (used on restore so a
-    /// persisted hash that no longer has a backing file falls back to no avatar).
-    func reloadPictureHashFromStore() {
-        WorkspacePictureStore.shared.invalidateCache(for: id)
-        pictureHash = WorkspacePictureStore.shared.pictureHash(for: id)
     }
 
     func setTerminalScrollBarHidden(_ hidden: Bool) {
